refactor: split jobs-utils into pure helpers + form-fields modules
Address architecture review feedback: separate pure formatting/status
helpers (jobs-helpers.ts) from form components (jobs-form-fields.tsx)
so the chart subtree doesn't depend on form-related imports.
