+++ apps/web/src/components/app/jobs-form-fields.tsx
@@ -1,0 +1,146 @@
+import { GitBranch } from "lucide-react";
+
+import { BranchSelect } from "@/components/app/branch-select";
+import { Checkbox } from "@/components/ui/checkbox";
+import { cn } from "@/lib/utils";
+
+export function SwitchToggle({
+  checked,
+  onCheckedChange,
+  ariaLabel,
+}: {
+  checked: boolean;
+  onCheckedChange: (checked: boolean) => void;
+  ariaLabel: string;
+}) {
+  return (
+    <button
+      type="button"
+      role="switch"
+      aria-checked={checked}
+      aria-label={ariaLabel}
+      onClick={() => onCheckedChange(!checked)}
+      className={cn(
+        "inline-flex h-6 w-11 shrink-0 items-center rounded-full border border-transparent p-0.5 transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 focus:ring-offset-background",
+        checked ? "bg-primary" : "bg-muted"
+      )}
+    >
+      <span
+        className={cn(
+          "h-5 w-5 rounded-full bg-background shadow-sm transition-transform",
+          checked ? "translate-x-5" : "translate-x-0"
+        )}
+      />
+    </button>
+  );
+}
+
+export function JobWorktreeOption({
+  checked,
+  cwd,
+  baseBranch,
+  branchName,
+  onCheckedChange,
+  onBaseBranchChange,
+  onBranchNameChange,
+  testIdPrefix,
+}: {
+  checked: boolean;
+  cwd: string;
+  baseBranch: string;
+  branchName: string;
+  onCheckedChange: (checked: boolean) => void;
+  onBaseBranchChange: (value: string) => void;
+  onBranchNameChange: (value: string) => void;
+  testIdPrefix?: string;
+}) {
+  return (
+    <div className="space-y-2 rounded-md border border-border/70 bg-muted/20 px-3 py-3 md:col-span-2">
+      <label className="flex cursor-pointer items-start gap-3">
+        <Checkbox
+          checked={checked}
+          onCheckedChange={() => onCheckedChange(!checked)}
+          className="mt-0.5"
+          title="Toggle git worktree"
+        />
+        <span className="space-y-1">
+          <span className="flex items-center gap-1.5 text-sm font-medium text-foreground">
+            <GitBranch className="h-3.5 w-3.5" />
+            Run in a git worktree
+          </span>
+          <span className="block text-xs text-muted-foreground">
+            Creates an isolated worktree and branch when this job runs.
+          </span>
+        </span>
+      </label>
+      {checked ? (
+        <div className="ml-8 w-[calc(100%-2rem)]">
+          <BranchSelect
+            cwd={cwd}
+            baseBranch={baseBranch}
+            onBaseBranchChange={onBaseBranchChange}
+            worktreeBranch={branchName}
+            onWorktreeBranchChange={onBranchNameChange}
+            testIdPrefix={testIdPrefix ?? "job-worktree"}
+          />
+        </div>
+      ) : null}
+    </div>
+  );
+}
+
+export function JobKeepAgentOption({
+  checked,
+  onCheckedChange,
+}: {
+  checked: boolean;
+  onCheckedChange: (checked: boolean) => void;
+}) {
+  return (
+    <label className="flex cursor-pointer items-start gap-3 rounded-md border border-border/70 bg-muted/20 px-3 py-3 md:col-span-2">
+      <Checkbox
+        checked={checked}
+        onCheckedChange={() => onCheckedChange(!checked)}
+        className="mt-0.5"
+        title="Keep agent after run completes"
+      />
+      <span className="space-y-1">
+        <span className="block text-sm font-medium text-foreground">
+          Keep agent after run completes
+        </span>
+        <span className="block text-xs text-muted-foreground">
+          The agent stays in your Agents list so you can continue the session
+          after the job finishes.
+        </span>
+      </span>
+    </label>
+  );
+}
+
+export function JobFullAccessOption({
+  checked,
+  onCheckedChange,
+}: {
+  checked: boolean;
+  onCheckedChange: (checked: boolean) => void;
+}) {
+  return (
+    <label className="flex cursor-pointer items-start gap-3 rounded-md border border-border/70 bg-muted/20 px-3 py-3 md:col-span-2">
+      <Checkbox
+        checked={checked}
+        onCheckedChange={() => onCheckedChange(!checked)}
+        className="mt-0.5"
+        title="Toggle full access"
+      />
+      <span className="space-y-1">
+        <span className="block text-sm font-medium text-foreground">
+          Run in full access mode
+        </span>
+        <span className="block text-xs text-muted-foreground">
+          Starts the selected agent with its most permissive supported execution
+          mode.
+        </span>
+      </span>
+    </label>
+  );
+}
